refactor: pkg api
